requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.1.0
Pillow==10.2.0
numpy==1.26.0
gunicorn==21.2.0